css-logical-properties-1/logicalprops-block-size.html

<!doctype html>
<meta charset="utf-8">
<title>CSS Logical Properties: {max-,min-}block-size</title>
<link rel="author" title="Xu Xing" href="mailto:openxu@gmail.com">
<link rel="help" href="https://drafts.csswg.org/css-logical-props-1/#logical-dimension-properties">
<link rel="help" href="https://drafts.csswg.org/css-writing-modes-3/#logical-to-physical">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/style-check.js"></script>

<style>
div {
  border: 1px solid #000;
}
#div1 {
  block-size: 40px;
  min-block-size: 50px;
  max-block-size: 100px;
}
#div2 {
  block-size: 100px;
  min-block-size: 50px;
  max-block-size: 100px;
}
#div3 {
  block-size: 120px;
  min-block-size: 50px;
  max-block-size: 100px;
}
#ref_div1 {
  height: 40px;
  min-height: 50px;
  max-height: 100px;
}
#ref_div2 {
  height: 100px;
  min-height: 50px;
  max-height: 100px;
}
#ref_div3 {
  height: 120px;
  min-height: 50px;
  max-height: 100px;
}

p {
  border: 1px solid #000;
}
#p1 {
  block-size: 100px;
  height: 50px;
}
#p2 {
  height: 50px;
  block-size: 100px;
}
#ref_p1 {
  height: 50px;
}
#ref_p2 {
  height: 100px;
}

.table {
  border: 1px solid #000;
  display: table;
}
.tablecolumn {
  display: table-cell;
}
#table1_column {
  block-size: 40px;
  min-block-size: 50px;
  max-block-size: 100px;
  inline-size: 100px;
  background-color: red;
}
#table2_column {
  block-size: 100px;
  min-block-size: 50px;
  max-block-size: 100px;
  inline-size: 100px;
  background-color: blue;
}
#table3_column {
  block-size: 120px;
  min-block-size: 50px;
  max-block-size: 100px;
  inline-size: 100px;
  background-color: green;
}
#ref_table1_column {
  height: 40px;
  min-height: 50px;
  max-height: 100px;
  width: 100px;
  background-color: red;
}
#ref_table2_column {
  height: 100px;
  min-height: 50px;
  max-height: 100px;
  width: 100px;
  background-color: blue;
}
#ref_table3_column {
  height: 120px;
  min-height: 50px;
  max-height: 100px;
  width: 100px;
  background-color: green;
}
</style>

<div id="div1">Hello, this is a test</div>
<div id="div2">Hello, this is a test</div>
<div id="div3">Hello, this is a test</div>
<div id="ref_div1">Hello, this is a test</div>
<div id="ref_div2">Hello, this is a test</div>
<div id="ref_div3">Hello, this is a test</div>

<p id="p1">Hello, this is a test</div>
<p id="p2">Hello, this is a test</div>
<p id="ref_p1">Hello, this is a test</div>
<p id="ref_p2">Hello, this is a test</div>

<div class="table">
  <div id="table1_column"></div>
</div>
<div class="table">
  <div id="table2_column"></div>
</div>
<div class="table">
  <div id="table3_column"></div>
</div>
<div class="table">
  <div id="ref_table1_column"></div>
</div>
<div class="table">
  <div id="ref_table2_column"></div>
</div>
<div class="table">
  <div id="ref_table3_column"></div>
</div>

<script>
test(function () {
  assert_true(compareWidthHeight("div1","ref_div1"));
  assert_true(compareWidthHeight("div2","ref_div2"));
  assert_true(compareWidthHeight("div3","ref_div3"));
}, "Check that block-size < min-block-size or min-block-size< block-size <=max-block-size or block-size > max-block-size");

test(function () {
  assert_true(compareWidthHeight("p1","ref_p1"));
  assert_true(compareWidthHeight("p2","ref_p2"));
}, "Check that height override block-size and vice versa");

test(function () {
  assert_true(compareWidthHeight("table1_column","ref_table1_column"));
  assert_true(compareWidthHeight("table2_column","ref_table2_column"));
  assert_true(compareWidthHeight("table3_column","ref_table3_column"));
}, "Check that block-size < min-block-size or min-block-size< block-size <=max-block-size or block-size > max-block-size in table");
</script>
